requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.34.0
aiogram>=3.17.0
google-genai>=1.0.0
yt-dlp>=2025.1.0
Pillow>=11.0.0
httpx>=0.28.0
python-dotenv>=1.0.0
python-multipart>=0.0.20
jinja2>=3.1.0
aiofiles>=24.1.0
APScheduler>=3.10.0
google-api-python-client>=2.0.0
google-auth-oauthlib>=1.0.0
